start splitting modal code into another file
